Done feature/login
When done fill out to id and password, to be click the login button.

and check on server, enter to next if correct user.
