+++ app/src/main/java/me/manage_outlet/MainActivity.kt
@@ -1,10 +1,14 @@
 package me.manage_outlet
 
+import android.content.Intent
 import android.os.AsyncTask
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
+import android.text.Editable
+import android.text.TextWatcher
 import android.util.Log
 import android.widget.EditText
+import android.widget.TextView
 import kotlinx.android.synthetic.main.activity_main.*
 import org.json.JSONObject
 import java.io.*
@@ -16,72 +20,52 @@ class MainActivity : AppCompatActivity(){
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        btnLogin.setOnClickListener{LoginTask(editId).execute()}
-    }
-}
-
-class LoginTask(val edt: EditText) : AsyncTask<String, Void, String>(){
-    override fun doInBackground(vararg p0: String?): String {
-        val url = URL("https://get-and-post-as-json.herokuapp.com/")
-        val httpClient = url.openConnection() as HttpURLConnection
-        var result = ""
-        try {
-            //Configure
-            httpClient.requestMethod = "POST"
-            httpClient.setRequestProperty("Cache-Control", "no-cache")
-            httpClient.setRequestProperty("Content-Type", "application/json")
-            httpClient.setRequestProperty("Accept", "text/html")
-//            httpClient.doInput = true
-            httpClient.doOutput = true
-
-            Log.d("check", "Try connect to Server...")
-            httpClient.connect()
-
-            //httpClient.outputStream is mean "Create Stream".
-            writeStream(outputStream = httpClient.outputStream)
-            result = readStream(inputStream = httpClient.inputStream)
-//            httpClient.responseCode
-
-            Log.d("check","Try outputStream.close() .")
-        }catch (e: Exception){
-            e.printStackTrace()
-            Log.d("check", "EXCEPTION : " + e.printStackTrace().toString())
-        }finally {
-            httpClient.disconnect()
-            Log.d("check", "Disconnected on Server.")
-        }
-        return result
-    }
-    fun readStream(inputStream: InputStream): String{
-        val bufferedReader: BufferedReader = BufferedReader(InputStreamReader(inputStream))
-        val stringBuffer = StringBuffer()
-        var responeResult : String? = ""
-
-        while(true) {
-            responeResult = bufferedReader.readLine()
-            if(responeResult != null){
-                stringBuffer.append(responeResult)
-            }else{
-                break
+        var checkFillout = arrayOf(0, 0)
+        btnLogin.isEnabled = false
+        editId.addTextChangedListener(object:TextWatcher{
+            override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
+                when(editId.text.toString().length){
+                    0 -> checkFillout[0] = 0
+                    else -> checkFillout[0] = 1
+                }
+                when(checkFillout[0] == 1 && checkFillout[1] == 1){
+                    true -> btnLogin.isEnabled = true
+                    else -> btnLogin.isEnabled = false
+                }
+            }
+            override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {}
+            override fun afterTextChanged(p0: Editable?) {}
+        })
+        editPw.addTextChangedListener(object:TextWatcher{
+            override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
+                when(editPw.text.toString().length){
+                    0 -> checkFillout[1] = 0
+                    else -> checkFillout[1] = 1
+                }
+                when(checkFillout[0] == 1 && checkFillout[1] == 1){
+                    true -> btnLogin.isEnabled = true
+                    else -> btnLogin.isEnabled = false
+                }
+            }
+            override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {}
+            override fun afterTextChanged(p0: Editable?) {}
+        })
+        btnLogin.setOnClickListener{
+            val localUrl = "https://6dcbc392.ngrok.io/login"
+            val herokuUrl = "https://get-and-post-as-json.herokuapp.com/login"
+            val result = LoginTask(editId, editPw, textMsg, this).execute(localUrl).get()
+            when(result){
+                "Login" -> loginWelcome()
+                "Wrong id" -> textMsg.text = result
+                "Wrong password" -> textMsg.text = result
+                else -> textMsg.text = "What..."
             }
         }
-        return stringBuffer.toString()
     }
-    fun writeStream(outputStream: OutputStream){
-        Log.d("check", "Entered in writeStream().")
-        val jsonObject = JSONObject()
-        jsonObject.put("id",edt.text.toString())
-        val bufferedWriter = BufferedWriter(OutputStreamWriter(outputStream))
-        Log.d("check","Try send the " + jsonObject.toString() + "to Server...")
-        bufferedWriter.write(jsonObject.toString())
-        Log.d("check","Try outputStream.flush() .")
-        bufferedWriter.flush()
-        bufferedWriter.close()
-    }
-
-    override fun onPostExecute(result: String?) {
-        super.onPostExecute(result)
-        Log.d("check","result: " + result)
-        edt.setText(result)
+    fun loginWelcome(){
+        textMsg.text = "Welcome"
+        val intent = Intent(this, ManageActivity::class.java)
+        startActivity(intent)
     }
 }
+
